Add setting the training attribute of children modules to equal the parent's after building
The problem is that, by default, new modules are created with the training attribute set to true and this might cause undesired normalization statistics updates or other state changes when the module is being built using an initial input for shape inference. Now this should not happen if the parent modules are in evaluation mode.